refactor toTitleCase, extract webhooks and calbacks
